updated project
The recent global health crisis also known as the COVID-19 or coronavirus pandemic has attracted the researchers' attentions to a treatment approach called immune plasma or convalescent plasma once more again. The main idea lying behind the immune plasma treatment is transferring the antibody rich part of the blood taken from the patients who are recovered previously to the critical individuals and its effciency has been proven by successfully using against great influenza of 1918, H1N1 flu, MERS, SARS and Ebola. In this study, we modeled the mentioned treatment approach and introduced a new meta-heuristic called Immune Plasma (IP) algorithm. The performance of the IP algorithm was investigated in detail and then compared with some of the classical and state-of-art meta-heuristics by solving a set of numerical benchmark problems. Moreover, the capabilities of the IP algorithm were also analyzed over complex engineering optimization problems related with the noise minimization of the electro-encephalography signal measurements. The results of the experimental studies showed that the IP algorithm is capable of obtaining better solutions for the vast majority of the test problems compared to other commonly used meta-heuristic algorithms.